Fixed RVM prompt and database.yml copy.
Prompt to see if using RVM.
Could not use Thor copy_file to create a database.yml backup.
By default using copy_file with a relative path will use the
"source path", rails_root, as the directory to find the it's source
and destination.
Needed to construct a full path to the generated app as the source and
destination path.